Fix pay dialog state being kept despite dialog closing
